pg-logic\Passwords\Loader\WordDictionaryLoader - Move all non-loader related methods to DictionaryPasswordGenerator, return the word tree instead of storing it (#11 #13).
