+++ pg-logic/Passwords/Generators/DictionaryPasswordGenerator.cs
@@ -23,9 +23,11 @@ namespace PG.Logic.Passwords.Generators
 		protected override bool RemoveHighAsciiCharacters => _options.RemoveHighAsciiCharacters;
 		protected override KeystrokeOrder KeystrokeOrder => _options.KeystrokeOrder;
 
+		private WordDictionaryTree? _wordTree;
+
 		public override GenerationResult Generate()
 		{
-			_dictionary.Load();
+			_wordTree = _dictionary.Load();
 
 			return base.Generate();
 		}
@@ -144,7 +146,7 @@ namespace PG.Logic.Passwords.Generators
 					_random.DiscardEntropy();
 					word = GenerateWord(averageLength, depthLevel, ref currentHand);
 				}
-				while (iterations++ < Constants.MAX_ITERATIONS && string.IsNullOrEmpty(word) && _dictionary.IsLeafNodeReached(word));
+				while (iterations++ < Constants.MAX_ITERATIONS && string.IsNullOrEmpty(word) && IsLeafNodeReached(word));
 
 				if (iterations >= Constants.MAX_ITERATIONS)
 					throw new InvalidOperationException("Max iterations reached without being able to generate a valid word.");
@@ -155,6 +157,50 @@ namespace PG.Logic.Passwords.Generators
 
 				currentHand = ChooseHand(currentHand);
 			}
+		}
+
+		/// <summary>
+		/// Searches for a leaf node in the dictionary tree by traversing the tree using the specified word and returns true if the leaf node is reached; 
+		/// the word is found.
+		/// </summary>
+		public bool IsLeafNodeReached(string word) => TrySearchLeafNode(word, out _);
+
+		/// <summary>
+		/// Searches for a leaf node in the dictionary tree by traversing the tree using the specified word. If the word is not found, the search stops at 
+		/// the last node that was found and returns false.
+		/// </summary>
+		private bool TrySearchLeafNode(string word, out ITreeNode<string> node)
+		{
+			node = _wordTree?.Root
+				?? throw new InvalidOperationException("The word tree has not been initialized.");
+
+			foreach (var letter in word)
+			{
+				var children = node.Children.Select(kvp => kvp.Value)
+					.FirstOrDefault(tn => tn.Value.ToString().Equals(letter.ToString(), StringComparison.InvariantCultureIgnoreCase));
+
+				if (children == default) return false;
+
+				node = children;
+			}
+
+			return true;
+		}
+
+		/// <summary>
+		/// Searches for the last possible leaf node in the dictionary tree by successively removing the last character of the word. If there is no valid 
+		/// node, the search stops at the last node that was found and returns false.
+		/// </summary>
+		public bool TrySearchLastPossibleLeafNode(string word, int depthLevel, out ITreeNode<string> node)
+		{
+			if (depthLevel <= 0)
+				throw new ArgumentOutOfRangeException(nameof(depthLevel), "Depth level must be greater than zero.");
+
+			bool found;
+			do { found = TrySearchLeafNode(word.Right(depthLevel--), out node); }
+			while (!found && depthLevel > 0);
+
+			return found;
 		}
 
 		protected override HandSide ChooseHand(HandSide currentHand)
@@ -181,7 +227,9 @@ namespace PG.Logic.Passwords.Generators
 			int wordLength = averageLength + (_random.Next(wordLengthVariance * 2) - wordLengthVariance);
 
 			Finger? curFinger = null;
-			ITreeNode<string> node = _dictionary.WordTree.Root;
+			ITreeNode<string> node = _wordTree?.Root
+				?? throw new InvalidOperationException("The word tree has not been initialized.");
+
 			foreach (int _ in Enumerable.Range(0, wordLength))
 			{
 				HandSide curHand = currentHand;
@@ -204,7 +252,7 @@ namespace PG.Logic.Passwords.Generators
 				node = next;
 
 				// If the word is already in the dictionary, break the loop.
-				if (wordBuilder.Length >= depthLevel && !_dictionary.TrySearchLastPossibleLeafNode(wordBuilder.ToString(), depthLevel, out node))
+				if (wordBuilder.Length >= depthLevel && !TrySearchLastPossibleLeafNode(wordBuilder.ToString(), depthLevel, out node))
 					break;
 			}
