+++ pg-logic/Passwords/Loader/IDictionaryLoader.cs
@@ -4,12 +4,6 @@ namespace PG.Logic.Passwords.Loader
 {
 	public interface IDictionaryLoader
 	{
-		internal WordDictionaryTree WordTree { get; set; }
-
-		void Load();
-
-		bool IsLeafNodeReached(string word);
-
-		bool TrySearchLastPossibleLeafNode(string word, int depthLevel, out ITreeNode<string> node);
+		WordDictionaryTree Load();
 	}
 }
